+++ docs/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixes
 - `scaffold_dict = resolve_scaffold(ORDER_OF_RESPECT, FILENAMES_OF_RESPECT)` added to the top of the `seed` CLI command due to failure of the seed command when no scaffold_dict was known. 
+
+### Added
+- --edit flag to `mulch seed`, to launch editor
+
 --
 
 ## [0.2.4] - 2025-07-11